+++ test/libsolidity/smtCheckerTests/functions/functions_trivial_condition_while.sol
@@ -1,8 +1,0 @@
-contract C
-{
-	function f(bool x) public pure { require(x); while (x) {} }
-}
-// ====
-// SMTEngine: all
-// ----
-// Warning 6838: (66-67): BMC: Condition is always true.