+++ test/libsolidity/smtCheckerTests/loops/do_while_bmc_iterations_continue_1.sol
@@ -1,0 +1,19 @@
+contract C {
+	function f() public pure {
+		uint x = 0;
+		do {
+			++x;
+			if (x == 3) {
+				continue;
+			}
+			++x;
+		} while (x < 3);
+		assert(x == 3);
+	}
+}
+// ====
+// SMTEngine: bmc
+// SMTSolvers: z3
+// BMCLoopIterations: 4
+// ----
+// Info 6002: BMC: 3 verification condition(s) proved safe! Enable the model checker option "show proved safe" to see all of them.