+++ test/libsolidity/smtCheckerTests/loops/for_loop_bmc_iterations_4.sol
@@ -1,0 +1,15 @@
+contract C
+{
+	function f() public pure {
+		uint x;
+		for (uint i = 0; i < 0; ++i)
+			++x;
+		assert(x == 0);
+	}
+}
+// ====
+// SMTEngine: bmc
+// SMTSolvers: z3
+// BMCLoopIterations: 3
+// ----
+// Info 6002: BMC: 3 verification condition(s) proved safe! Enable the model checker option "show proved safe" to see all of them.