+++ test/libsolidity/smtCheckerTests/loops/for_loop_bmc_iterations_break_3.sol
@@ -1,0 +1,18 @@
+contract C
+{
+	function f() public pure {
+		uint x;
+		for (uint i = 0; i < 3; ++i) {
+			if (i > 0)
+				break;
+			++x;
+		}
+		assert(x == 1);
+	}
+}
+// ====
+// SMTEngine: bmc
+// SMTSolvers: z3
+// BMCLoopIterations: 4
+// ----
+// Info 6002: BMC: 3 verification condition(s) proved safe! Enable the model checker option "show proved safe" to see all of them.