+++ test/libsolidity/smtCheckerTests/loops/while_break_direct.sol
@@ -1,5 +1,4 @@
-contract C
-{
+contract C {
 	function f(uint x) public pure {
 		x = 0;
 		while (x < 10)
@@ -12,4 +11,3 @@ contract C
 // SMTSolvers: z3
 // ----
 // Info 1391: CHC: 1 verification condition(s) proved safe! Enable the model checker option "show proved safe" to see all of them.
-// Warning 6838: (65-71): BMC: Condition is always true.